lab sheet components separated. Laboratory projects have been added to the home page.
